requirements.txt
celluloid==0.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
deepdish==0.3.7
easing-functions==1.0.4
et-xmlfile==1.1.0
fonttools==4.33.3
h5py==3.7.0
idna==3.3
imageio==2.19.3
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
networkx==2.8.2
nibabel==3.2.2
nilearn==0.9.1
nltools==0.4.5
numexpr==2.8.1
numpy==1.22.4
opencv-contrib-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
py-feat==0.3.7
pynv==0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
tables==3.7.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9